requirements.txt
streamlit
openai
pdfplumber
python-dotenv
pytesseract
pillow
PyMuPDF